refactor: split DirectoryBrowser into smaller components
- Created shared types in types.ts
- Split DirectoryList into its own component
- Created NewDirectoryDialog component
- Extracted logic into useDirectoryBrowser hook
- Updated main DirectoryBrowser to use new components
